+++ Frends.RabbitMQ.Publish/CHANGELOG.md
@@ -1,7 +1,10 @@
-# Changelog
+﻿# Changelog
 
 ## [1.7.0] - 2025-02-14
-### Fixed
+### Added
+- New property: Result.Success.
+- New feature: Added certificate‑based authentication, including new required connection properties. 
+### Changed
 - Update RabbitMQ.Client from 7.0.0 to 7.2.0
 - Update System.Runtime.Caching from 9.0.2 to 10.0.2
 
@@ -11,7 +14,7 @@
 - Fix the issue with binding unnamed queue.
 
 ## [1.5.0] - 2025-02-14
-### Fixed
+### Changed
 - Update RabbitMQ.Client from 6.8.1 to 7.0.0
 - Update System.Runtime.Caching from 8.0.0 to 9.0.2